fix(tasks): scope task queries by campaign id, clamp copies, guard loading state
Review follow-ups on the DM task settings feature:
- SessionTask model/controller take req.campaignId explicitly on every query
  as defence in depth alongside RLS (058 seeded every campaign).
- Tasks page clamps a task's copies to the phase headcount so nobody draws
  the same task twice, and distinguishes "still loading" / "failed to load"
  from "no tasks defined" before assigning.
- Migration 058: NULL::smallint cast keeps the VALUES list type-stable.

+++ backend/migrations/058_add_session_task_definitions.sql
@@ -74,7 +74,7 @@ INSERT INTO session_task_definition (campaign_id, phase, name, quantity, min_cha
 SELECT c.id, d.phase, d.name, d.quantity, d.min_characters, d.is_snack_master, d.sort_order
 FROM campaigns c
 CROSS JOIN (VALUES
-    ('pre',    'Get Dice Trays',                                          1, NULL, false, 1),
+    ('pre',    'Get Dice Trays',                                          1, NULL::smallint, false, 1),
     ('pre',    'Put Initiative name tags on tracker',                     1, NULL, false, 2),
     ('pre',    'Wipe TV',                                                 1, NULL, false, 3),
     ('pre',    'Recap',                                                   1, NULL, false, 4),
